Use renderables instead of console writers

+++ PwshSpectreConsole/Build.ps1
@@ -1,5 +1,7 @@
 param (
-    [string] $Version = "0.49.0"
+    [string] $Version = "0.49.1",
+    [int] $DotnetSdkMajorVersion = 8,
+    [switch] $NoReinstall
 )
 
 function Install-SpectreConsole {
@@ -53,9 +55,14 @@ function Install-SpectreConsole {
 
     $command = Get-Command "dotnet" -ErrorAction SilentlyContinue
     if ($null -eq $command) {
-        throw "dotnet not found, please install dotnet sdk 6"
-    } elseif (-not (dotnet --list-sdks | Select-String "^6.+")) {
-        throw "dotnet sdk 6 not found, please install dotnet sdk 6"
+        throw "dotnet not found, please install dotnet sdk $DotnetSdkMajorVersion"
+    } elseif (-not (dotnet --list-sdks | Select-String "^$DotnetSdkMajorVersion.+")) {
+        Write-Warning "dotnet sdk $DotnetSdkMajorVersion not found, please install dotnet sdk $DotnetSdkMajorVersion"
+        if (Get-Command "winget" -ErrorAction SilentlyContinue) {
+            winget install "Microsoft.DotNet.SDK.$DotnetSdkMajorVersion"
+        } else {
+            throw "Please install the dotnet sdk and try again"
+        }
     }
     try {
         Push-Location
@@ -70,6 +77,12 @@ Write-Host "Downloading Spectre.Console version $Version"
 $installLocation = (Join-Path $PSScriptRoot "packages")
 $csharpProjectLocation = (Join-Path $PSScriptRoot "private" "classes")
 $testingInstallLocation = (Join-Path $PSScriptRoot ".." "PwshSpectreConsole.Tests" "packages")
+
+if ((Test-Path $installLocation) -or (Test-Path $testingInstallLocation) -and $NoReinstall) {
+    Write-Host "Spectre.Console already installed, skipping"
+    return
+} 
+
 if (Test-Path $installLocation) {
     Remove-Item $installLocation -Recurse -Force
 }
